# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Avido Experiments MVP Mock API",
    "description": "Mock API for the Experiments feature in Avido. Provides endpoints for experiments, variants, steps, tasks, tests, and evaluations. All endpoints require an orgId query parameter for tenant isolation.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev server"
    }
  ],
  "paths": {
    "/api/experiments": {
      "get": {
        "summary": "List experiments",
        "description": "Returns experiments for the specified organization with optional filtering and sorting.",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "description": "Organization ID (required for security)",
            "required": true,
            "schema": { "type": "string" },
            "example": "org-avido-demo"
          },
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search across name and description",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by experiment status",
            "required": false,
            "schema": { "$ref": "#/components/schemas/ExperimentStatus" }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["name", "createdAt", "status"]
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of experiments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Experiment" }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - orgId missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{id}/variants": {
      "get": {
        "summary": "List variants for an experiment",
        "description": "Returns variants for a specific experiment with computed metrics including test results and eval scores.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Experiment ID",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "orgId",
            "in": "query",
            "description": "Organization ID (required for security)",
            "required": true,
            "schema": { "type": "string" },
            "example": "org-avido-demo"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by variant status",
            "required": false,
            "schema": { "$ref": "#/components/schemas/VariantStatus" }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["name", "createdAt", "status"]
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of variants with metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VariantWithMetrics" }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - orgId missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Access denied - orgId mismatch",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Experiment not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/steps": {
      "get": {
        "summary": "List steps (LLM keys)",
        "description": "Returns available steps (LLM keys) that can be configured in experiments.",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "description": "Organization ID (required for security)",
            "required": true,
            "schema": { "type": "string" },
            "example": "org-avido-demo"
          },
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search across name, externalId, and description",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by step type",
            "required": false,
            "schema": { "$ref": "#/components/schemas/StepType" }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["name", "externalId", "type"]
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of steps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Step" }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - orgId missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/evals": {
      "get": {
        "summary": "List evaluations",
        "description": "Returns evaluations with optional filtering and the ability to include eval definition metadata.",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "description": "Organization ID (required for security)",
            "required": true,
            "schema": { "type": "string" },
            "example": "org-avido-demo"
          },
          {
            "name": "testId",
            "in": "query",
            "description": "Filter by test ID",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "definitionId",
            "in": "query",
            "description": "Filter by eval definition ID",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by evaluation status",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TestStatus" }
          },
          {
            "name": "passed",
            "in": "query",
            "description": "Filter by pass/fail status",
            "required": false,
            "schema": { "type": "boolean" }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Search eval definitions by name",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "createdAtFrom",
            "in": "query",
            "description": "Filter evals created at or after this timestamp",
            "required": false,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "createdAtTo",
            "in": "query",
            "description": "Filter evals created before or at this timestamp",
            "required": false,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "scoreMin",
            "in": "query",
            "description": "Filter by minimum score",
            "required": false,
            "schema": { "type": "number" }
          },
          {
            "name": "scoreMax",
            "in": "query",
            "description": "Filter by maximum score",
            "required": false,
            "schema": { "type": "number" }
          },
          {
            "name": "includeDefinition",
            "in": "query",
            "description": "Include eval definition metadata in response",
            "required": false,
            "schema": { "type": "boolean", "default": false }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["createdAt", "score", "status", "id"]
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort direction",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of evaluations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Eval" }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - orgId missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Error message" }
        }
      },
      "ExperimentStatus": {
        "type": "string",
        "enum": ["DRAFT", "BASELINE_PENDING", "BASELINE_RUNNING", "BASELINE_COMPLETED", "RUNNING", "COMPLETED", "ARCHIVED"],
        "description": "Status of an experiment"
      },
      "VariantStatus": {
        "type": "string",
        "enum": ["DRAFT", "RUNNING", "COMPLETED", "FAILED", "REJECTED"],
        "description": "Status of a variant"
      },
      "TestStatus": {
        "type": "string",
        "enum": ["PENDING", "IN_PROGRESS", "FAILED", "COMPLETED"],
        "description": "Status of a test execution or evaluation"
      },
      "StepType": {
        "type": "string",
        "enum": ["LLM", "RETRIEVAL", "TOOL", "PROCESSING"],
        "description": "Type of step in the application"
      },
...
      "Experiment": {
        "type": "object",
        "required": ["id", "orgId", "name", "status", "stepIds", "taskIds", "createdBy", "createdAt", "modifiedAt"],
        "properties": {
          "id": { "type": "string", "description": "Unique identifier" },
          "orgId": { "type": "string", "description": "Organization ID" },
          "name": { "type": "string", "description": "Experiment name" },
          "description": { "type": "string", "description": "Optional description" },
          "status": { "$ref": "#/components/schemas/ExperimentStatus" },
          "stepIds": { "type": "array", "items": { "type": "string" }, "description": "Steps in scope" },
          "taskIds": { "type": "array", "items": { "type": "string" }, "description": "Fixed task set" },
          "createdBy": { "type": "string", "description": "User who created the experiment" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ConfigPatch": {
        "type": "object",
        "description": "Configuration overrides for a variant",
        "properties": {
          "system_prompt": { "type": "string" },
          "temperature": { "type": "number" },
          "top_p": { "type": "number" },
          "max_tokens": { "type": "integer" },
          "model": { "type": "string" }
        }
      },
      "Variant": {
        "type": "object",
        "required": ["id", "orgId", "experimentId", "name", "status", "configPatch", "createdBy", "createdAt", "modifiedAt"],
        "properties": {
          "id": { "type": "string", "description": "Unique identifier" },
          "orgId": { "type": "string", "description": "Organization ID" },
          "experimentId": { "type": "string", "description": "Parent experiment ID" },
          "previousVariantId": { "type": "string", "description": "Variant this was inherited from" },
          "name": { "type": "string", "description": "Variant name" },
          "status": { "$ref": "#/components/schemas/VariantStatus" },
          "targetStepId": { "type": "string", "description": "Step being modified (null for baseline)" },
          "configPatch": { "$ref": "#/components/schemas/ConfigPatch" },
          "description": { "type": "string", "description": "Optional description" },
          "createdBy": { "type": "string", "description": "User who created the variant" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" }
        }
      },
      "VariantMetrics": {
        "type": "object",
        "required": ["totalTests", "completedTests", "failedTests", "avgScore", "passRate", "evalBreakdown"],
        "properties": {
          "totalTests": { "type": "integer", "description": "Total number of tests" },
          "completedTests": { "type": "integer", "description": "Number of completed tests" },
          "failedTests": { "type": "integer", "description": "Number of failed tests" },
          "avgScore": { "type": "number", "description": "Average eval score" },
          "passRate": { "type": "number", "description": "Percentage of evals that passed" },
          "evalBreakdown": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "definitionId": { "type": "string" },
                "name": { "type": "string" },
                "avgScore": { "type": "number" },
                "passRate": { "type": "number" }
              }
            }
          }
        }
      },
      "VariantWithMetrics": {
        "allOf": [
          { "$ref": "#/components/schemas/Variant" },
          {
            "type": "object",
            "properties": {
              "metrics": { "$ref": "#/components/schemas/VariantMetrics" }
            }
          }
        ]
      },
      "Step": {
        "type": "object",
        "required": ["id", "orgId", "externalId", "name", "type", "createdAt", "modifiedAt"],
        "properties": {
          "id": { "type": "string", "description": "Unique identifier" },
          "orgId": { "type": "string", "description": "Organization ID" },
          "externalId": { "type": "string", "description": "External identifier used in webhooks" },
          "name": { "type": "string", "description": "Human-readable name" },
          "description": { "type": "string", "description": "Optional description" },
          "type": { "$ref": "#/components/schemas/StepType" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Eval": {
        "type": "object",
        "required": ["id", "orgId", "applicationId", "testId", "definitionId", "status", "results", "score", "passed", "scoreComputedAt", "createdAt", "modifiedAt"],
        "properties": {
          "id": { "type": "string", "description": "Unique identifier" },
          "orgId": { "type": "string", "description": "Organization ID" },
          "applicationId": { "type": "string", "description": "Application ID" },
          "testId": { "type": "string", "description": "Test ID" },
          "definitionId": { "type": "string", "description": "Eval definition ID" },
          "status": { "$ref": "#/components/schemas/TestStatus" },
          "results": { "type": "object", "description": "Eval results (type varies by eval type)" },
          "score": { "type": "number", "description": "Computed score (0-1 for most types, 0-5 for naturalness)" },
          "passed": { "type": "boolean", "description": "Whether the eval passed" },
          "scoreComputedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}